got the rendering to work inside the clay layout
